CircularQueue renamed, refactored according to inspection hints.
